init.lua: Implement replace_in_buffer function
This function will be used to update the content of the buffer after
creating a new abbreviation, so the user will not need to manually fix
his typos after adding them to the correction dictionary.

This functionality is enabled by default, but can be disabled with
relevant options for the setup function.