init.lua: Load dictionaries sequentially
The plugin is working with 2 different dictionaries, a global
dictionary, and a local dictionary. When the plugins starts, it loads
both dictionaries.

As these dictionaries (especially the global one) can be very big,
loading them happens asynchronously, to not block the start of the
nvim.

The original implementation added the 2 dictionary loading to the main
loop, without waiting or synchronizing them, making the loading order of
their abbreviates inconsistent.
In case there are entries which are the same in both dicts, this makes
the created abbreviation inconsistent.

This commit solves this problem by making the dictionary loading
sequential - first, the code loads the local user dictionary, and only
once finished, it loads the global dictionary.
This means that if both dictionaries contain the same typo, the value
from the local dictionary takes precedence (as the code adds new values
to the abbreviates only if they do not exist in the dict yet).

The implication of loading the dictionaries sequentially is that now it
takes these dictionaries slightly longer to load (as previously they
were loaded at the same time, while now the global dictionary starts
loading only after the local one is finished). However, assuming a
reasonable size for the local dictionary, this should not cause any
issue.

The sequential loading implementation is somewhat cumbersome. This is
done as nvim-lua does not create a consistent way to wait for functions
that starts with `vim.defer_fn`, which requires this commit to locally
implement such a function, and update the calling function with a way to
signal finishing to the user.
As the deferred function cannot return a value to signal finishing, this
is done by passing a table to the function, that is changed when the
function has finished.