init.lua: Don't add empty abbreviates
An easy way to implement local removal of values from the global
dictionary is by creating an abbreviate of "typo" -> "typo" in the local
dictionary.
This will now work, as the local dictionary takes precedence over the
global one (see the previous commit).

To make this feature more efficient, the code is updated to keep storing
this value as a "typo" (so the global value will not be loaded), but not
creating any abbreviation for this value, as this abbreviation is not
needed.