init.lua: Serialize abbrev removal differently
As per the previous commit, an abbreviation removal is defined as a
key-value pair, where the key and the value are the same.

This commit updates the MistakeEdit buffer with a new syntax for
removing an abbreviation. When an abbreviation is marked as removed,
instead of having the syntax of "typo" -> "typo" (which will still
work), the plugin will show this abbreviation as "!typo", for easier
editing of this file.