OMSA: README.md: Update repository references
Update the references to the repository to be to omrisarig13, which is
the current fork, instead of to the base project. This is done as work
on this fork is done locally, and the README can reflect that better
here.

This commit should not be pushed/merged to the base repository.