OMSA: mistake.lua: Delete file
In my setup, I'm using lazy.nvim to load the plugin. This makes this
file unnecessary - actually, when this file exists, it re-runs the setup
function, which will ignore any different default configuration that is
set in the personal config file.

I should re-investigated this commit, understand if this is a
me-problem, or if there is something bigger here, that the base
repository can also benefit from.